ensure storage upload path exists on init, delete old temp files every hour, add storage delete methods, convert url to file for uploader, allow existing value in uploader, display video preview in uploader

+++ packages/server/src/storage/tmp.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, Logger, NotFoundException, OnModuleInit } from '@nestjs/common';
-import * as fg from 'fast-glob';
+import { Cron, CronExpression } from '@nestjs/schedule';
 import * as fs from 'fs-extra';
 
 export const TEMP_PATH = '/tmp/uploads';
@@ -13,6 +13,33 @@ export class TmpService implements OnModuleInit {
    */
   async onModuleInit() {
     await fs.ensureDir(TEMP_PATH);
+  }
+
+  /**
+   * Remove old files from temp path every hour
+   */
+  @Cron(CronExpression.EVERY_HOUR)
+  async removeOldFiles() {
+    this.logger.verbose('📂 Checking for old files to remove from temp path');
+
+    // Get all files in the temp path that are older than an hour
+    const files = await fs.readdir(TEMP_PATH);
+    // Set cutoff to be an hour ago
+    const livesUntil = Date.now() - 3600000;
+    let count = 0;
+
+    for (const file of files) {
+      const path = `${TEMP_PATH}/${file}`;
+      const stats = await fs.stat(path);
+      // Check if the file is older than the cutoff
+      if (stats.mtimeMs < livesUntil) {
+        this.logger.verbose(`📂 Removing old file ${path}`);
+        await fs.remove(path);
+        count++;
+      }
+    }
+
+    this.logger.verbose(`📂 Removed ${count} old files from temp path`);
   }
 
   /**
@@ -42,14 +69,15 @@ export class TmpService implements OnModuleInit {
    */
   async deleteFile(fileName: string): Promise<void> {
     this.logger.verbose(`📂 Deleting file "${fileName}" from temp path`);
+    const path = `${TEMP_PATH}/${fileName}`;
 
-    // Get all files that match the file name
-    const paths = await fg([`${TEMP_PATH}/${fileName}*`]);
+    // Check that the file exists
+    if (!(await fs.stat(path))) {
+      throw new NotFoundException(`File "${fileName}" not found`);
+    }
 
     // Delete the file
-    for (const path of paths) {
-      fs.remove(path);
-    }
+    fs.remove(path);
     this.logger.debug(`✅️ Deleted file "${fileName}" from temp path`);
   }
 }
